+++ providers/implementations/ciphers/cipher_aes_ccm.h
@@ -56,7 +56,7 @@ int ossl_cipher_set_ccm_aes_initkey(PROV_CCM_CTX *ctx,
 const PROV_CCM_HW *ossl_prov_aes_hw_ccm(size_t keylen);
 
 #if defined(AESNI_CAPABLE)
-const PROV_CCM_HW *ossl_prov_aes_hw_ccm_aesni(size_t keybits);
+const PROV_CCM_HW *ossl_prov_aes_hw_ccm_aesni(void);
 #endif
 #if defined(OPENSSL_CPUID_OBJ) && defined(__riscv) && __riscv_xlen == 32
 const PROV_CCM_HW *ossl_prov_aes_hw_ccm_rv32i(size_t keybits);
