+++ providers/implementations/ciphers/cipher_aes_gcm_hw.c
@@ -156,7 +156,7 @@ const PROV_GCM_HW *ossl_prov_aes_hw_gcm(size_t keybits)
     const PROV_GCM_HW *aes_gcm_hw = NULL;
 
 #if defined(AESNI_CAPABLE)
-    aes_gcm_hw = ossl_prov_aes_hw_gcm_aesni(keybits);
+    aes_gcm_hw = ossl_prov_aes_hw_gcm_aesni();
 #elif defined(AES_PMULL_CAPABLE) && defined(AES_GCM_ASM)
     aes_gcm_hw = ossl_prov_aes_hw_gcm_armv8(keybits);
 #elif defined(PPC_AES_GCM_CAPABLE) && defined(_ARCH_PPC64)
