+++ providers/implementations/ciphers/cipher_aes_hw.c
@@ -217,7 +217,7 @@ static const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_mode(enum aes_modes mode,
     const PROV_CIPHER_HW *aes_hw_mode = NULL;
 
 #if defined(AESNI_CAPABLE)
-    aes_hw_mode = ossl_prov_cipher_hw_aesni(mode, keybits);
+    aes_hw_mode = ossl_prov_cipher_hw_aesni(mode);
 #elif defined(ARMv8_HWAES_CAPABLE)
     aes_hw_mode = ossl_prov_cipher_hw_arm(mode, keybits);
 #elif defined(OPENSSL_CPUID_OBJ) && defined(__riscv) && __riscv_xlen == 32
