+++ providers/implementations/ciphers/cipher_aes_xts_hw.c
@@ -15,7 +15,7 @@
 
 #include "cipher_aes_xts.h"
 
-static int cipher_set_aes_xts_initkey(PROV_CIPHER_CTX *ctx,
+int ossl_cipher_set_aes_xts_initkey(PROV_CIPHER_CTX *ctx,
     const unsigned char *key, size_t keylen,
     aes_set_encrypt_key_fn fn_set_enc_key,
     aes_set_encrypt_key_fn fn_set_dec_key,
@@ -62,7 +62,7 @@ static int cipher_hw_aes_xts_generic_initkey(PROV_CIPHER_CTX *ctx,
 #ifdef HWAES_xts_decrypt
         stream_dec = HWAES_xts_decrypt;
 #endif /* HWAES_xts_decrypt */
-        return cipher_set_aes_xts_initkey(ctx, key, keylen,
+        return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
             HWAES_set_encrypt_key, HWAES_set_decrypt_key,
             HWAES_encrypt, HWAES_decrypt, stream_enc, stream_dec);
     }
@@ -72,7 +72,7 @@ static int cipher_hw_aes_xts_generic_initkey(PROV_CIPHER_CTX *ctx,
     if (BSAES_CAPABLE) {
         stream_enc = ossl_bsaes_xts_encrypt;
         stream_dec = ossl_bsaes_xts_decrypt;
-        return cipher_set_aes_xts_initkey(ctx, key, keylen,
+        return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
             AES_set_encrypt_key, AES_set_decrypt_key,
             AES_encrypt, AES_decrypt, stream_enc, stream_dec);
     }
@@ -80,18 +80,18 @@ static int cipher_hw_aes_xts_generic_initkey(PROV_CIPHER_CTX *ctx,
 
 #ifdef VPAES_CAPABLE
     if (VPAES_CAPABLE) {
-        return cipher_set_aes_xts_initkey(ctx, key, keylen,
+        return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
             vpaes_set_encrypt_key, vpaes_set_decrypt_key,
             vpaes_encrypt, vpaes_decrypt, stream_enc, stream_dec);
     }
 #endif /* VPAES_CAPABLE */
 
-    return cipher_set_aes_xts_initkey(ctx, key, keylen,
+    return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
         AES_set_encrypt_key, AES_set_decrypt_key,
         AES_encrypt, AES_decrypt, stream_enc, stream_dec);
 }
 
-static void cipher_hw_aes_xts_copyctx(PROV_CIPHER_CTX *dst,
+void ossl_cipher_hw_aes_xts_copyctx(PROV_CIPHER_CTX *dst,
     const PROV_CIPHER_CTX *src)
 {
     PROV_AES_XTS_CTX *sctx = (PROV_AES_XTS_CTX *)src;
@@ -102,56 +102,7 @@ static void cipher_hw_aes_xts_copyctx(PROV_CIPHER_CTX *dst,
     dctx->xts.key2 = &dctx->ks2.ks;
 }
 
-#if defined(AESNI_CAPABLE)
-
-static int cipher_hw_aesni_xts_initkey(PROV_CIPHER_CTX *ctx,
-    const unsigned char *key, size_t keylen)
-{
-    void (*aesni_xts_enc)(const unsigned char *in,
-        unsigned char *out,
-        size_t length,
-        const AES_KEY *key1, const AES_KEY *key2,
-        const unsigned char iv[16]);
-    void (*aesni_xts_dec)(const unsigned char *in,
-        unsigned char *out,
-        size_t length,
-        const AES_KEY *key1, const AES_KEY *key2,
-        const unsigned char iv[16]);
-
-    aesni_xts_enc = aesni_xts_encrypt;
-    aesni_xts_dec = aesni_xts_decrypt;
-
-#if (defined(__x86_64) || defined(__x86_64__) || defined(_M_AMD64) || defined(_M_X64))
-    if (aesni_xts_avx512_eligible()) {
-        if (keylen == 64) {
-            aesni_xts_enc = aesni_xts_256_encrypt_avx512;
-            aesni_xts_dec = aesni_xts_256_decrypt_avx512;
-        } else if (keylen == 32) {
-            aesni_xts_enc = aesni_xts_128_encrypt_avx512;
-            aesni_xts_dec = aesni_xts_128_decrypt_avx512;
-        }
-    }
-#endif
-
-    return cipher_set_aes_xts_initkey(ctx, key, keylen,
-        aesni_set_encrypt_key, aesni_set_decrypt_key,
-        aesni_encrypt, aesni_decrypt, aesni_xts_enc, aesni_xts_dec);
-}
-
-static const PROV_CIPHER_HW aesni_xts = {
-    cipher_hw_aesni_xts_initkey,
-    NULL,
-    cipher_hw_aes_xts_copyctx
-};
-
-static const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_xts_aesni()
-{
-    if (AESNI_CAPABLE)
-        return &aesni_xts;
-    return NULL;
-}
-
-#elif defined(SPARC_AES_CAPABLE)
+#if defined(SPARC_AES_CAPABLE)
 
 static int cipher_hw_aes_xts_t4_initkey(PROV_CIPHER_CTX *ctx,
     const unsigned char *key, size_t keylen)
@@ -173,7 +124,7 @@ static int cipher_hw_aes_xts_t4_initkey(PROV_CIPHER_CTX *ctx,
         return 0;
     }
 
-    return cipher_set_aes_xts_initkey(ctx, key, keylen,
+    return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
         aes_t4_set_encrypt_key, aes_t4_set_decrypt_key,
         aes_t4_encrypt, aes_t4_decrypt, stream_enc, stream_dec);
 }
@@ -181,7 +132,7 @@ static int cipher_hw_aes_xts_t4_initkey(PROV_CIPHER_CTX *ctx,
 static const PROV_CIPHER_HW aes_xts_t4 = {
     cipher_hw_aes_xts_t4_initkey,
     NULL,
-    cipher_hw_aes_xts_copyctx
+    ossl_cipher_hw_aes_xts_copyctx
 };
 
 static const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_xts_t4()
@@ -199,13 +150,13 @@ static int cipher_hw_aes_xts_rv64i_initkey(PROV_CIPHER_CTX *ctx,
     if (RISCV_HAS_ZVBB() && RISCV_HAS_ZVKG() && RISCV_HAS_ZVKNED() && riscv_vlen() >= 128) {
         /* Zvkned only supports 128 and 256 bit keys. */
         if (keylen * 8 == 128 * 2 || keylen * 8 == 256 * 2)
-            return cipher_set_aes_xts_initkey(ctx, key, keylen,
+            return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
                 rv64i_zvkned_set_encrypt_key, rv64i_zvkned_set_decrypt_key,
                 rv64i_zvkned_encrypt, rv64i_zvkned_decrypt,
                 rv64i_zvbb_zvkg_zvkned_aes_xts_encrypt,
                 rv64i_zvbb_zvkg_zvkned_aes_xts_decrypt);
 
-        return cipher_set_aes_xts_initkey(ctx, key, keylen,
+        return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
             AES_set_encrypt_key, AES_set_encrypt_key,
             rv64i_zvkned_encrypt, rv64i_zvkned_decrypt, NULL, NULL);
     }
@@ -213,17 +164,17 @@ static int cipher_hw_aes_xts_rv64i_initkey(PROV_CIPHER_CTX *ctx,
     if (RISCV_HAS_ZVKNED() && riscv_vlen() >= 128) {
         /* Zvkned only supports 128 and 256 bit keys. */
         if (keylen * 8 == 128 * 2 || keylen * 8 == 256 * 2)
-            return cipher_set_aes_xts_initkey(ctx, key, keylen,
+            return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
                 rv64i_zvkned_set_encrypt_key, rv64i_zvkned_set_decrypt_key,
                 rv64i_zvkned_encrypt, rv64i_zvkned_decrypt, NULL, NULL);
 
-        return cipher_set_aes_xts_initkey(ctx, key, keylen,
+        return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
             AES_set_encrypt_key, AES_set_encrypt_key,
             rv64i_zvkned_encrypt, rv64i_zvkned_decrypt, NULL, NULL);
     }
 
     if (RISCV_HAS_ZKND_AND_ZKNE())
-        return cipher_set_aes_xts_initkey(ctx, key, keylen,
+        return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
             rv64i_zkne_set_encrypt_key, rv64i_zknd_set_decrypt_key,
             rv64i_zkne_encrypt, rv64i_zknd_decrypt, NULL, NULL);
 
@@ -233,7 +184,7 @@ static int cipher_hw_aes_xts_rv64i_initkey(PROV_CIPHER_CTX *ctx,
 static const PROV_CIPHER_HW aes_xts_rv64i = {
     cipher_hw_aes_xts_rv64i_initkey,
     NULL,
-    cipher_hw_aes_xts_copyctx
+    ossl_cipher_hw_aes_xts_copyctx
 };
 
 static const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_xts_rv64i()
@@ -250,13 +201,13 @@ static int cipher_hw_aes_xts_rv32i_initkey(PROV_CIPHER_CTX *ctx,
     const unsigned char *key, size_t keylen)
 {
     if (RISCV_HAS_ZBKB_AND_ZKND_AND_ZKNE())
-        return cipher_set_aes_xts_initkey(ctx, key, keylen,
+        return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
             rv32i_zbkb_zkne_set_encrypt_key,
             rv32i_zbkb_zknd_zkne_set_decrypt_key,
             rv32i_zkne_encrypt, rv32i_zknd_decrypt, NULL, NULL);
 
     if (RISCV_HAS_ZKND_AND_ZKNE())
-        return cipher_set_aes_xts_initkey(ctx, key, keylen,
+        return ossl_cipher_set_aes_xts_initkey(ctx, key, keylen,
             rv32i_zkne_set_encrypt_key, rv32i_zknd_zkne_set_decrypt_key,
             rv32i_zkne_encrypt, rv32i_zknd_decrypt, NULL, NULL);
 
@@ -266,7 +217,7 @@ static int cipher_hw_aes_xts_rv32i_initkey(PROV_CIPHER_CTX *ctx,
 static const PROV_CIPHER_HW aes_xts_rv32i = {
     cipher_hw_aes_xts_rv32i_initkey,
     NULL,
-    cipher_hw_aes_xts_copyctx
+    ossl_cipher_hw_aes_xts_copyctx
 };
 
 static const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_xts_rv32i()
@@ -419,7 +370,7 @@ static const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_xts_s390x(size_t keybits)
 static const PROV_CIPHER_HW aes_generic_xts = {
     cipher_hw_aes_xts_generic_initkey,
     NULL,
-    cipher_hw_aes_xts_copyctx
+    ossl_cipher_hw_aes_xts_copyctx
 };
 
 const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_xts(size_t keybits)
